2001-02-01 Bill Tutt <rassilon@lima.mudlib.org>     * SVN.cpp (CSVNWorkingCopy::wc_statuses): Add cast to remove     compiler warning.     * MarshalArray.h: Document limitation with current code not following all of the connection point rules atm.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@838456 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/clients/win32/svn_com/SVN.cpp
@@ -343,7 +343,7 @@ CSVNWorkingCopy::wc_statuses(BSTR bstrPath, SAFEARRAY **ppsa)
 	SafeArrayUnaccessData(psa);
 	fLockedSA = FALSE;
 
-	if (rgsBound.cElements > i) {
+	if (rgsBound.cElements > (UINT)i) {
 		rgsBound.cElements = i;
 		hr = SafeArrayRedim(psa, &rgsBound);
 		if (FAILED(hr))
